+++ src/rest_framework_reactive/connection.py
@@ -8,10 +8,10 @@ def get_queryobserver_settings():
         'warnings': {'max_result_length': 1000, 'max_processing_time': 1.0},
         # Observers going over these limits will be stopped.
         'errors': {'max_processing_time': 20.0},
-        # Update batch delay (in seconds). If a new update comes earlier than the
-        # delay value, queue processing will be delayed so multiple updates can be
-        # batched. A higher value introduces more latency.
-        'update_batch_delay': 5,
+        # Throttle evaluation (in seconds). If a new update comes earlier than
+        # given rate value, the evaluation will be delayed (and batched).
+        # A higher value introduces more latency.
+        'throttle_rate': 2,
     }
     defaults.update(getattr(settings, 'DJANGO_REST_FRAMEWORK_REACTIVE', {}))
     return defaults