+++ src/rest_framework_reactive/consumers.py
@@ -6,12 +6,44 @@ from asgiref.sync import async_to_sync
 from channels.consumer import AsyncConsumer, SyncConsumer
 from channels.generic.websocket import JsonWebsocketConsumer
 
+from django.core.cache import cache
+
+from .connection import get_queryobserver_settings
 from .models import Observer, Subscriber
 from .observer import QueryObserver
 from .protocol import *
 
 # Maximum number of cached observer executors.
 MAX_CACHED_EXECUTORS = 1024
+# Throttle constants
+THROTTLE_CACHE_PREFIX = 'drf_reactive_observer_throttle_'
+THROTTLE_SEMAPHORE_DELAY = 'delay'
+THROTTLE_SEMAPHORE_EVALUATE = 'evaluate'
+THROTTLE_SEMAPHORE_IGNORE = 'ignore'
+
+
+def throttle_cache_key(observer_id):
+    return '{}{}'.format(THROTTLE_CACHE_PREFIX, observer_id)
+
+
+def throttle_semaphore(observer_id):
+    """Check if observer should be evaluated, delayed or ignored.
+
+    Increase the observer counter if throttle cache exists.
+    """
+    cache_key = throttle_cache_key(observer_id)
+    throttle_rate = get_queryobserver_settings()['throttle_rate']
+
+    try:
+        count = cache.incr(cache_key)
+        # Ignore if delayed observer already scheduled.
+        return THROTTLE_SEMAPHORE_DELAY if count == 2 else THROTTLE_SEMAPHORE_IGNORE
+    except ValueError:
+        count = cache.get_or_set(cache_key, default=1, timeout=throttle_rate)
+        # Ignore if cache was set and increased in another thread.
+        return THROTTLE_SEMAPHORE_EVALUATE if count == 1 else THROTTLE_SEMAPHORE_IGNORE
+
+    assert False  # This should never happen.
 
 
 class PollObserversConsumer(AsyncConsumer):
@@ -26,6 +58,24 @@ class PollObserversConsumer(AsyncConsumer):
         await self.channel_layer.send(
             CHANNEL_WORKER_NOTIFY,
             {'type': TYPE_EVALUATE_OBSERVER, 'observer': message['observer']},
+        )
+
+
+class ThrottleConsumer(AsyncConsumer):
+    """Consumer for throttling observer evaluations."""
+
+    async def delay_observer_evaluate(self, message):
+        """Throttle observer evaluation."""
+        observer_id = message['observer']
+        throttle_rate = get_queryobserver_settings()['throttle_rate']
+
+        # Sleep until we need to schedule the next evaluation.
+        await asyncio.sleep(throttle_rate)
+
+        # Dispatch task to evaluate the observable.
+        await self.channel_layer.send(
+            CHANNEL_WORKER_NOTIFY,
+            {'type': TYPE_EVALUATE_OBSERVER, 'observer': observer_id},
         )
 
 
@@ -69,22 +119,36 @@ class WorkerConsumer(SyncConsumer):
             .values_list('pk', flat=True)
         )
 
-        for observer in observers:
+        for observer_id in observers:
             async_to_sync(self.channel_layer.send)(
                 CHANNEL_WORKER_NOTIFY,
-                {'type': TYPE_EVALUATE_OBSERVER, 'observer': observer},
+                {'type': TYPE_EVALUATE_OBSERVER, 'observer': observer_id},
             )
 
     def observer_evaluate(self, message):
         """Evaluate observer."""
+        observer_id = message['observer']
+        throttle_rate = get_queryobserver_settings()['throttle_rate']
+
         try:
-            observer = Observer.objects.only('pk', 'request').get(
-                pk=message['observer']
-            )
+            observer = Observer.objects.only('pk', 'request').get(pk=observer_id)
         except Observer.DoesNotExist:
             return
 
-        self._evaluate(observer)
+        if throttle_rate <= 0:
+            # Evaluate observer.
+            self._evaluate(observer)
+            return
+
+        action = throttle_semaphore(observer_id)
+        if action == THROTTLE_SEMAPHORE_EVALUATE:
+            # Evaluate observer.
+            self._evaluate(observer)
+        elif action == THROTTLE_SEMAPHORE_DELAY:
+            # Delay observer by throttle rate.
+            async_to_sync(self.channel_layer.send)(
+                CHANNEL_THROTTLE, {'type': TYPE_THROTTLE, 'observer': observer_id}
+            )
 
 
 class ClientConsumer(JsonWebsocketConsumer):